CMakePM: Treat .h files as compile group language header types
CMake splits sources files into groups "Source Files" and "Header
Files".

CMake also has compiler groups when source files are compiled
differently.

Qt Creator is mapping the compiler groups as RawProjectParts.

In order to get the header files as part of a RawProjectPart the target
sources (which contains all sources) is mapping the header files that
match the mime type of the compiler group language type.

.h header files were considered ambigous headers, and in this
commit we treat them as the compile group language header.

Fixes: QTCREATORBUG-27117
Change-Id: If68e847846cc270f06fc2231ec44a29ea6a987c1
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>