McuSupport: add template for qmlproject-based projects
Adds a new option for creating qmlproject-based MCU projects
in the "New Project" wizard. The generated qmlproject file
contains all the supported properties and nodes, prefilled
with their default values, for easy tweaking by users without
requiring to dive into the documentation.

No extra fonts, translations or images are generated, but
empty placeholder nodes for them are found in the qmlproject file,
for the user to fill in.
One custom module with a simple Text item is generated, in
order to demonstrate how to use the modules feature.

The required QmlProjectExporter version for building the project
is the one packaged with Qt for MCUs 2.3.0.

Task-number: QTCREATORBUG-26042
Change-Id: I4569306f2f6e7b9370dfa65cacd0b65c2e4fcb06
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Piotr Mućko <piotr.mucko@qt.io>
Reviewed-by: Dawid Śliwa <dawid.sliwa@qt.io>
Reviewed-by: Erik Verbruggen <erik.verbruggen@me.com>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>