Dumper: Fix dump of std::string on macOS
Try to handle the alternate layout of strings correctly.
Depending on the defines and the endianness the string structure
may vary quite a bit. Old approach just took care of the default
layout and the endianness.
In case of an alternate layout of strings we failed so far.

Fixes: QTCREATORBUG-26175
Done-with: Viktor Govako
Change-Id: I788eb5619408bca281eb887c3f6a269808c27d24
Reviewed-by: hjk <hjk@qt.io>