kitdetector: Fix detection on macOS
On macOS calling clang sometimes produces some output to stderr.
This would be captured by "runGcc" when detecting the triplet, which
would then show up in the preferences and had negative effects
on abi detection later on. To work around this, we only capture
the first line of output.

On macOS the auto detection code incorrectly assumed that if the
architecture is different it would not be able to run on the host.
This is not true if the Host is Arm, and the target is x86.

Change-Id: I3cd151dc6ad83142fb9643ba7b03a155e754c6d0
Reviewed-by: Eike Ziller <eike.ziller@qt.io>