QtSupport: Try harder at providing the right debugger
... for a temporary kit created by the importer.
The way the importer works is that it sets up a random kit (with a
default toolchain) and then overwrites toolchain and Qt version. If the
new toolchain is of a different type than the original one, the debugger
will not match anymore. Fix this by calling fix() on the changed kit and
adapting the DebuggerKitAspect so that it sets itself up from scratch if
the current value doesn't match the toolchain.

Fixes: QTCREATORBUG-27758
Change-Id: Ib0539f07c0c38728de89a49299b7d2439afc5744
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: hjk <hjk@qt.io>