Fix display of build errors/warnings in status bar
We show a summary of errors/warnings with icons in the status bar, if
the "Progress Details" are toggled off. The change that fixed the size
of that details widget to a grid, to prevent constant relayouting of the
output pane buttons while indexing runs, broke the size of the build
result summary.

The size was adapted in
ProgressManagerPrivate::updateStatusDetailsWidget() but that is only
called when FutureProgress::setWidget is called. When the build manager
later makes the warning/error details visible, the size was left at the
minimum size of 20 pixels, leading to the widget being cut off.

Amends b4f2ac0dd45c62d04a75c3e445734a0b37961dec

Change-Id: Ia41cf3db7e4485dff4712412976c1d4f5dd315b2
Reviewed-by: David Schulz <david.schulz@qt.io>