Fix the runtime warning when building against Qt 5
It fixes the following warning:
QObject::connect: Cannot queue arguments of type 'FilePath'
(Make sure 'FilePath' is registered using qRegisterMetaType().)

It looks like with Qt 5 all arguments needs to be specified
with full namespace in order to be used in queued connections.
Note, that warning is misleading as adding qRegisterMetaType()
doesn't fix it in this case.

The full scope isn't needed when building against Qt 6,
so leaving a TODO for Creator 9.0

Change-Id: I43a52385b8d3bb0ef90ce274e11df9fd68e62109
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: hjk <hjk@qt.io>