DockWidget: Avoid using sender()
The toggleView() was used as a signal handler for
m_toggleViewAction and as a public method. In the latter
case checking the sender doesn't make any sense.
Add a lambda as a direct signal handler for m_toggleViewAction
and perform a custom modification of the open arg before
calling toggleView().

Simplify setToggleViewActionChecked() implementation.

Change-Id: If3799110b3f1dcf6fe0c653ebcfd23f9716cb2f3
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Henning Gründl <henning.gruendl@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>