Fix regression with permissions when running app locally
Amends 3efa0f3961f53adda53157915a82707048acd3a6

The code path for running locally in some cases modifies the environment
(to support SUDO_ASKPASS), or the command (to support privacy settings
on macOS), but this was overwritten again after the above change.

This lead to e.g. the Qt camera example just crashing on macOS, instead
of the OS asking the user for permission to use the camera & microphone.

Change-Id: I41192bac736e4b07eb3530b37e35ad02a79e20a5
Reviewed-by: hjk <hjk@qt.io>