BuildEnvironmentWidget: Override 'Open Terminal' func to open build dir
If the user is opening a terminal in the build environment then probably
the intention is to do something in the build dir rather than in the home dir.

Also, most shells/terminals have a shortcut to go to the home dir but they
have no idea about the build dir so it is much harder to get there.

Change-Id: Ia259def7a4915fe86370fc183fb1e38ef6644c4d
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>