Fix a wrong property conversion
QVariant::canConvert returns just the general ability to
convert from one type to the other and does not reflect
if the specific string inside the variant can be converted.
For this reason tmp.canConvert(QMetaType::Double) returned
always true which lead to colors inside keyframes becoming
black as soon as the scene was saved.

Fixes: QDS-6960
Change-Id: I1a267d93a7cffb080ae884df490e5723592a2780
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Aleksei German <aleksei.german@qt.io>