jsonwizard: Fix check for icon
The jsonwizardfactory did not check if the template contained
an icon. Due to a bug in FilePath::resolvePath() which previously
returned "/" instead of "" for empty paths, the "iconPath.exists()"
check would always return "true".

With this change, the existence of the icon is only checked, if
an icon was supplied by the template in the first place.

Change-Id: I3252a6b92b34433835ab0e6e04857114ea895b46
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>