External Tools: Force console output for Qt based tools
Depending on platform and configuration and where they are run from
(terminal or not), Qt applications can send some output to a different
output stream than stdout/err (like Windows debug stream, or Linux
journald). This is a bit confusing because e.g. we capture that kind of
output for user applications run in Qt Creator. By default prevent it
with the corresponding environment variable. Can still be overridden by
setting QT_LOGGING_TO_CONSOLE=0 if that really poses a problem.

Fixes: QTCREATORBUG-27828
Change-Id: I7e3f9930b045fc887144753384ba246ecdc52d90
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>