Auto-detect ccache compilers on macOS
From some standard locations. When installing ccache with Homebrew, it
is installed in architecture dependent directory and links are
automatically created. When installing ccache with MacPorts, no links
are created, but the MacPorts wiki at trac.macports.org has a suggestion
for a script that creates them at /opt/local/libexec/ccache in its howto
section.

Fixes: QTCREATORBUG-27792
Change-Id: Ib173f81384a03b867cb54f45dac698ead95ecef6
Reviewed-by: Cristian Adam <cristian.adam@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>