Take "Toggle Progress Details" immediately into effect
When the progress details are visible, and the user clicks the button,
the details should immediately vanish. Before this patch, the progress
details stay visible until the user moves the mouse away from the
button, so clicking the button doesn't have an immediate effect, which
is pretty confusing.

Task-number: QTCREATORBUG-28078
Change-Id: Ia80ab48fc0a012e8ae17dea454f83747e5717e72
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>