QmlDesigner: Fix TextField context menu
Fix TextField keeping its selection after context menu is opened. Due to
this bug QTBUG-71723 being closed the behavior changed. Also the states
when condition TextField needed to be context menu aware to not set a
new expression when the context menu is opened, otherwise it will
immediately reset the model and the context menu wont show up if the
content was edited.

Task-number: QDS-7730
Change-Id: Ic0e63cd3d244bd7f83bcb0edd3b31c0ca1f0a5df
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>