Core: Go to pattern start when pressing home in locator
If the cursor is already at that position move the cursor to the start
of the line edit. This mimics the editor behavior that home moves the
cursor to the relevant position for that line, and if the cursor is at
that position move to the actual start of the line.

Change-Id: Ieedbc0de0f93b0ffa3f3b64ccbcb04471b004e8e
Reviewed-by: Eike Ziller <eike.ziller@qt.io>