QmlDesigner: Fix crash
Do not call validId in updateContext.

Calling validId does mutate the model and will set an id if none exists.
Since updateContext is called on selection change the user does not
expect the model to mutate and when switching/creating documents this
could even lead to a crash.

Instead, we give the created MosueArea always an id and disable the menu
in case there is no id.

Task-number: QDS-7806
Change-Id: I3d5d1ce41025d94db3cfff48f483366e56b58a32
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>