ClearCasePlugin: Introduce runCleartoolProc
Don't use VcsCommand when the only passed flag is NoOutput
as in this case VcsCommand is the same as QtcProcess.
Add runCleartoolProc and implement it using QtcProcess.
Use it whenever NoOutput has been passed to runCommand.
Some codepath, like runCleartoolSync() or ccGetActiveVobs()
has been called from non-GUI thread - be on the safe side
and use QtcProcess in these cases.

Change-Id: I2d29947393bd43af193a53e3f8e89e1b6b4b86fa
Reviewed-by: Orgad Shaneh <orgads@gmail.com>